Add parameter gitlab_unicorn_port to bind Unicorn to a specific port
When running Gitlab behind an SSL terminating Apache webserver, there
is a chance that port 8080 is already in use locally. This is also
the standard port for Unicorn in the Gitlab omnibus package. This
patch adds the ability to change the Unicorn port in Ansible using
the "gitlab_unicorn_port" parameter.