apply gitlab application settings
This commit introduces application settings for gitlab role.
It comes with 2 new parameters:

gitlab_apply_applications_settings: on/off feature to apply settings
gitlab_application_settings: a dictionary of the application settings
that may be applied.

All settings are only applied when neccessary. This is done
by first fetching the settings using gitlab-rails runner
and return the in json format.

To lookup all possible application settings have a look
at the command that fetches the possible settings.

A convenient step is added to transform python boolean
in to string lowercase values that can be used when
new settings are applied using the RoR ActiveRecord.update method.